refactor : 불필요한 이미지파일 제거
